flask_server/requirements.txt
flask
flask_cors
neo4j
psycopg2
python-dotenv